// assets/scripts/proto/proto_id_map.ts
// auto-generated by generate_proto_ids.py, do not edit

/** 协议名 -> numberkey */
export const PROTO_ID_BY_NAME = {
    CS_BagStoreDeposit: 53221,
    CS_BagUseItem: 99775,
    CS_BagWear: 37632,
    CS_ChangeCareer: 75508,
    CS_CheckCreateProhibit: 15788,
    CS_CheckReconnect: 92966,
    CS_CheckRoleName: 13878,
    CS_CreateRole: 81997,
    CS_EnterGame: 57740,
    CS_GetBagList: 40016,
    CS_GetPlayerAttr: 52982,
    CS_GetPlayerInfo: 17676,
    CS_IsHasRole: 89377,
    CS_Ping: 8275,
    CS_QuickLogin: 89258,
    CS_RandomRoleName: 44974,
    CS_Reconnect: 3892,
    CS_RenameRole: 80740,
    CS_SoulForge: 14143,
    CS_SwitchObserve: 69013,
    CS_VerifyAccount: 35505,
    SC_BagIdListNtf: 1475,
    SC_BagIdListNtf2: 80203,
    SC_BagItemListNtf: 69827,
    SC_BagItemListNtf2: 30513,
    SC_BagPageListNtf: 71907,
    SC_BagPartNtf: 74119,
    SC_BagPartNtf2: 35574,
    SC_BagSingleIdNtf: 56491,
    SC_BagStoreDeposit: 95375,
    SC_BagTipNtf: 56119,
    SC_BagUseItem: 50492,
    SC_BagWear: 40307,
    SC_ChangeCareer: 62743,
    SC_CheckCreateProhibit: 58952,
    SC_CheckReconnect: 74526,
    SC_CheckRoleName: 97766,
    SC_CreateRole: 87762,
    SC_EnterGame: 834,
    SC_GetBagList: 98148,
    SC_GetPlayerAttr: 41084,
    SC_GetPlayerInfo: 72635,
    SC_IsHasRole: 11166,
    SC_KickNotify: 22609,
    SC_MiscNotify: 60941,
    SC_Ping: 24230,
    SC_QuickLogin: 69682,
    SC_RandomRoleName: 51744,
    SC_Reconnect: 73793,
    SC_RenameRole: 1304,
    SC_SoulForge: 15952,
    SC_SwitchObserve: 45431,
    SC_TimeNotify: 91113,
    SC_VerifyAccount: 55360,
} as const;

/** numberkey -> 协议名 */
export const PROTO_NAME_BY_ID: Record<number, string> = {
    834: "SC_EnterGame",
    1304: "SC_RenameRole",
    1475: "SC_BagIdListNtf",
    3892: "CS_Reconnect",
    8275: "CS_Ping",
    11166: "SC_IsHasRole",
    13878: "CS_CheckRoleName",
    14143: "CS_SoulForge",
    15788: "CS_CheckCreateProhibit",
    15952: "SC_SoulForge",
    17676: "CS_GetPlayerInfo",
    22609: "SC_KickNotify",
    24230: "SC_Ping",
    30513: "SC_BagItemListNtf2",
    35505: "CS_VerifyAccount",
    35574: "SC_BagPartNtf2",
    37632: "CS_BagWear",
    40016: "CS_GetBagList",
    40307: "SC_BagWear",
    41084: "SC_GetPlayerAttr",
    44974: "CS_RandomRoleName",
    45431: "SC_SwitchObserve",
    50492: "SC_BagUseItem",
    51744: "SC_RandomRoleName",
    52982: "CS_GetPlayerAttr",
    53221: "CS_BagStoreDeposit",
    55360: "SC_VerifyAccount",
    56119: "SC_BagTipNtf",
    56491: "SC_BagSingleIdNtf",
    57740: "CS_EnterGame",
    58952: "SC_CheckCreateProhibit",
    60941: "SC_MiscNotify",
    62743: "SC_ChangeCareer",
    69013: "CS_SwitchObserve",
    69682: "SC_QuickLogin",
    69827: "SC_BagItemListNtf",
    71907: "SC_BagPageListNtf",
    72635: "SC_GetPlayerInfo",
    73793: "SC_Reconnect",
    74119: "SC_BagPartNtf",
    74526: "SC_CheckReconnect",
    75508: "CS_ChangeCareer",
    80203: "SC_BagIdListNtf2",
    80740: "CS_RenameRole",
    81997: "CS_CreateRole",
    87762: "SC_CreateRole",
    89258: "CS_QuickLogin",
    89377: "CS_IsHasRole",
    91113: "SC_TimeNotify",
    92966: "CS_CheckReconnect",
    95375: "SC_BagStoreDeposit",
    97766: "SC_CheckRoleName",
    98148: "SC_GetBagList",
    99775: "CS_BagUseItem",
};

export type ProtoMessageName = keyof typeof PROTO_ID_BY_NAME;